fix: rename storagev2 to storage
