+++ pkg/storage/batch.go
