+++ pkg/storage/inmemstore/inmembatch.go
