+++ pkg/storage/leveldbstore/batch.go
