--- pkg/storage/mock/storer_test.go
@@ -1,60 +1,0 @@
-// Copyright 2020 The Swarm Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-package mock_test
-
-import (
-	"bytes"
-	"context"
-	"errors"
-	"testing"
-
-	"github.com/ethersphere/bee/pkg/storage"
-	"github.com/ethersphere/bee/pkg/storage/mock"
-	"github.com/ethersphere/bee/pkg/swarm"
-)
-
-func TestMockStorer(t *testing.T) {
-	t.Parallel()
-
-	s := mock.NewStorer()
-
-	keyFound, err := swarm.ParseHexAddress("aabbcc")
-	if err != nil {
-		t.Fatal(err)
-	}
-	keyNotFound, err := swarm.ParseHexAddress("bbccdd")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	valueFound := []byte("data data data")
-
-	ctx := context.Background()
-	if _, err := s.Get(ctx, storage.ModeGetRequest, keyFound); !errors.Is(err, storage.ErrNotFound) {
-		t.Fatalf("expected ErrNotFound, got %v", err)
-	}
-
-	if _, err := s.Get(ctx, storage.ModeGetRequest, keyNotFound); !errors.Is(err, storage.ErrNotFound) {
-		t.Fatalf("expected ErrNotFound, got %v", err)
-	}
-
-	if _, err := s.Put(ctx, storage.ModePutUpload, swarm.NewChunk(keyFound, valueFound)); err != nil {
-		t.Fatalf("expected not error but got: %v", err)
-	}
-
-	if chunk, err := s.Get(ctx, storage.ModeGetRequest, keyFound); err != nil {
-		t.Fatalf("expected no error but got: %v", err)
-	} else if !bytes.Equal(chunk.Data(), valueFound) {
-		t.Fatalf("expected value %s but got %s", valueFound, chunk.Data())
-	}
-
-	has, err := s.Has(ctx, keyFound)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !has {
-		t.Fatal("expected mock store to have key")
-	}
-}
