+++ pkg/storage/storagetest/batch.go
